+++ src/Phan/Issue.php
@@ -309,6 +309,7 @@ class Issue
     public const ParamTooFew               = 'PhanParamTooFew';
     public const ParamTooFewInternal       = 'PhanParamTooFewInternal';
     public const ParamTooFewCallable       = 'PhanParamTooFewCallable';
+    public const ParamTooFewInPHPDoc       = 'PhanParamTooFewInPHPDoc';
     public const ParamTooMany              = 'PhanParamTooMany';
     public const ParamTooManyUnpack        = 'PhanParamTooManyUnpack';
     public const ParamTooManyInternal      = 'PhanParamTooManyInternal';
@@ -2923,6 +2924,14 @@ class Issue
                 "Call with {COUNT} arg(s) to {FUNCTIONLIKE} (as a provided callable) which requires {COUNT} arg(s) defined at {FILE}:{LINE}",
                 self::REMEDIATION_B,
                 7044
+            ),
+            new Issue(
+                self::ParamTooFewInPHPDoc,
+                self::CATEGORY_PARAMETER,
+                self::SEVERITY_LOW,
+                'Call with {COUNT} arg(s) to {FUNCTIONLIKE} which has phpdoc indicating it requires {COUNT} arg(s) (${PARAMETER} is mandatory) defined at {FILE}:{LINE}',
+                self::REMEDIATION_B,
+                7049
             ),
             new Issue(
                 self::ParamSpecial1,